refactor(svm-sdk): move prepareProgramUpgrade from warp/ to deploy/
The function is program-agnostic — only uses generic program bytes,
signer, RPC, and the deploy-plan utilities. Hoisting it lets the IGP
writer reuse it without depending on the warp module.
