+++ typescript/svm-sdk/src/warp/collateral-token.ts
@@ -35,7 +35,7 @@ import { hasProgramBytes } from '../types.js';
 import type { AnnotatedSvmTransaction, SvmReceipt, SvmRpc } from '../types.js';
 
 import type { SvmDeployedWarpAddress, SvmWarpTokenConfig } from './types.js';
-import { prepareProgramUpgrade } from './warp-upgrade.js';
+import { prepareProgramUpgrade } from '../deploy/program-upgrade.js';
 import {
   fetchCollateralTokenAccount,
   fetchWarpProgramVersion,
