+++ typescript/svm-sdk/src/warp/synthetic-token.ts
@@ -50,7 +50,7 @@ import type {
 } from '../types.js';
 
 import type { SvmDeployedWarpAddress, SvmWarpTokenConfig } from './types.js';
-import { prepareProgramUpgrade } from './warp-upgrade.js';
+import { prepareProgramUpgrade } from '../deploy/program-upgrade.js';
 import {
   fetchSyntheticTokenAccount,
   fetchWarpProgramVersion,
